requirements.txt
google-generativeai>=0.3.0
PyMuPDF>=1.23.0
requests>=2.31.0